requirements/local.txt
arrow==1.2.0
asgiref==3.4.1
autopep8==1.5.7
binaryornot==0.4.4
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.7
click==8.0.1
colorama==0.4.4
cookiecutter==1.7.3
cryptography==35.0.0
defusedxml==0.7.1
Django==3.2.8
django-allauth==0.45.0
django-appconf==1.0.5
django-compressor==2.4.1
django-crispy-forms==1.13.0
django-sass-processor==1.1
et-xmlfile==1.1.0
idna==3.3
Jinja2==3.0.2
jinja2-time==0.2.0
libsass==0.21.0
MarkupSafe==2.0.1
oauthlib==3.1.1
poyo==0.5.0
psycopg2-binary==2.9.1
pycodestyle==2.7.0
pycparser==2.20
PyJWT==2.2.0
python-dateutil==2.8.2
python-slugify==5.0.2
python3-openid==3.2.0
pytz==2021.3
rcssmin==1.0.6
requests==2.26.0
requests-oauthlib==1.3.0
rjsmin==1.1.0
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
toml==0.10.2
urllib3==1.26.7
